pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ekm"
version = "0.1.9"
description = "Episodic Knowledge Mesh - A novel knowledge management system with attention-driven sparse retrieval"
readme = "README.md"
authors = [{name = "Achref Riahi", email = "achrefr556@gmail.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "torch>=1.10.0",
    "faiss-cpu>=1.7.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "requests>=2.25.0",
    "aiohttp>=3.8.0",
    "openai>=1.0.0",
    "google-genai>=0.2.0",
    "tenacity>=8.0.0",
    "python-dotenv>=0.19.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.20.0",
    "matplotlib>=3.5.0",
    "pypdf>=3.0.0",
    "httpx>=0.24.0",
    "redis>=4.5.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.16.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
    "build",
    "wheel"
]
gpu = [
    "faiss-gpu>=1.7.0"
]
test = [
    "pytest>=6.0",
    "pytest-asyncio>=0.16.0",
    "pytest-cov>=3.0.0",
    "psutil>=5.8.0"
]

[project.urls]
Homepage = "https://github.com/kachrefv/ekm"
Repository = "https://github.com/kachrefv/ekm.git"
Issues = "https://github.com/kachrefv/ekm/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ekm"]
exclude = ["tests*", "examples*", "docs*", "notebooks*", "scripts*", "benchmarks*", "migrations*", "data*", "results*"]

[tool.setuptools.package-data]
ekm = ["*.md", "*.txt", "*.yaml", "*.yml"]